Added device name to description
Added the name of the device to the snapshot description. This may not always match exactly with the device in the OS (on CentOS it may show as /dev/xvdf but as /dev/sdf in the AWS console, for exmple) but should give enough context to be able to determine what the drive is.